# internal/transport/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HOLibrary API
  version: "1.0.0"
  description: |
    Online library management system for **Hezekiah Oluwasanmi Library**,
    Obafemi Awolowo University, Ile-Ife.

    Group 4 · SEN 106 / SEN 216, Introduction to Web Technologies

    ## What this API is

    A modernised Online Public Access Catalogue with an integrated circulation
    module, for a **physical** library. It answers *does HOL have this book, is a
    copy free, where is it* and records *who holds which copy, from when, until
    when*. No book content is stored or served.

    ## Two ideas that shape every endpoint

    **A book is not a copy.** A `book` is bibliographic, one title, one call
    number, shared by all its volumes. A `copy` is one physical object with its
    own **accession number**, unique across the whole collection. Five copies of
    *Clean Code* are one book and five copies.

    **Availability and overdue are derived, never stored.** There is no
    `available_copies` field in the database and no `is_overdue` column. Both are
    computed on read, so neither can drift out of step with reality.

    ## Membership begins in the building

    There is **no registration endpoint**, by design. At HOL a user presents an
    identity card and signs a register; a librarian then creates the account.
    The member receives a single-use password-setup link by email, and no
    password is sent or returned by the API.

    ## Time

...
    display only.

...

...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Sign out
      description: |
        Revokes the refresh token. The access token is left to expire on its own
        (15 minutes), the trade-off that comes with stateless access tokens, and
        the reason they are short-lived.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200": { $ref: '#/components/responses/StatusOK' }
        "401": { $ref: '#/components/responses/Unauthenticated' }
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change your own password
      description: |
        Requires the current password, so a borrowed unlocked laptop is not
        enough to take over an account.

        **Every other session for this account is revoked.** Changing a password
        is how someone reacts to a suspected compromise; if old refresh tokens
        kept working it would achieve nothing.

        This is the one route available to an account still holding a
        librarian-issued temporary password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password:
                  type: string
                  format: password
                  description: At least 10 characters, with a letter and a number.
      responses:
        "200": { $ref: '#/components/responses/StatusOK' }
        "401": { $ref: '#/components/responses/Unauthenticated' }
        "422": { $ref: '#/components/responses/PasswordTooWeak' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <access_token>` from `/auth/login`.

        Access tokens last 15 minutes. Refresh tokens last 7 days, are stored
        hashed, and are rotated on every use.
...
  responses:
    StatusOK:
      description: Done
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  status: { type: string }
    Unauthenticated:
      description: No valid access token (`UNAUTHENTICATED`, `TOKEN_INVALID`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PasswordTooWeak:
      description: At least 10 characters, with a letter and a number (`PASSWORD_TOO_WEAK`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable and machine-readable. Switch on this.
              example: COPY_NOT_AVAILABLE
            message:
              type: string
              description: Written for a person. Safe to display; reveals nothing internal.
              example: That copy is already on loan.
            details: { type: object, nullable: true }
...